Move HeterogeneousProductBase to its own header
