+++ HeterogeneousCore/Product/interface/HeterogeneousProduct.h
@@ -4,18 +4,14 @@
 #include "FWCore/Utilities/interface/Exception.h"
 
 #include "HeterogeneousCore/Product/interface/HeterogeneousDeviceId.h"
+#include "HeterogeneousCore/Product/interface/HeterogeneousProductBase.h"
 
-#include <bitset>
 #include <cassert>
 #include <functional>
 #include <memory>
-#include <mutex>
 #include <tuple>
 
 namespace heterogeneous {
-  constexpr const unsigned int kMaxDevices = 16;
-  using DeviceBitSet = std::bitset<kMaxDevices>;
-
   template <typename T>
   std::string bitsetArrayToString(const T& bitsetArray) {
     std::string ret;
@@ -195,32 +191,6 @@ namespace heterogeneous {
     using type = Empty;
   };
 }
-
-// For type erasure to ease dictionary generation
-class HeterogeneousProductBase {
-public:
-  // TODO: Given we'll likely have the data on one or at most a couple
-  // of devices, storing the information in a "dense" bit pattern may
-  // be overkill. Maybe a "sparse" presentation would be sufficient
-  // and easier to deal with?
-  using BitSet = heterogeneous::DeviceBitSet;
-  using BitSetArray = std::array<BitSet, static_cast<unsigned int>(HeterogeneousDevice::kSize)>;
-
-  virtual ~HeterogeneousProductBase() = 0;
-
-  bool isProductOn(HeterogeneousDevice loc) const {
-    // should this be protected with the mutex?
-    return location_[static_cast<unsigned int>(loc)].any();
-  }
-  BitSet onDevices(HeterogeneousDevice loc) const {
-    // should this be protected with the mutex?
-    return location_[static_cast<unsigned int>(loc)];
-  }
-
-protected:
-  mutable std::mutex mutex_;
-  mutable BitSetArray location_;
-};
 
 /**
  * Generic data product for holding data on CPU or a heterogeneous
